+++ html/elements/footer.php
@@ -1,8 +1,8 @@
-<div style="width: 1100px; margin: 0 auto;">
+<footer class="hbmon-shell hbmon-footer">
 <p style="text-align: center;"><span style="text-align: center;">
 Cortney T. Buffington, N0MJS. <a target="_blank" href="https://github.com/ShaYmez/hblink3">HBlink3.</a>  Copyright &copy; 2016-2025<br>The Regents of the <a target="_blank" href="http://k0usy.mystrikingly.com/">K0USY Group</a>. All Rights Reserved.<br><a title="HBMonv2 SP2ONG v20211012" target="_blank" href="https://github.com/sp2ong/HBMonv2">Original Version SP2ONG 2019-2025</a><br><a title="HBMonv2 ShaYmez" target="_blank" href="https://github.com/ShaYmez/HBMonv2">Dashboard Version <?php echo htmlspecialchars(defined('DASH') ? DASH : ''); ?></a><?php if (defined('DOCKER_VERSION') && DOCKER_VERSION !== '') { ?><br><a title="HBlink Docker by Shaymez" target="_blank" href="https://github.com/ShaYmez/hblink3-docker-install">Docker Version <?php echo htmlspecialchars(DOCKER_VERSION); ?></a><?php } ?><br></span>
     <!-- THIS COPYRIGHT NOTICE MUST BE DISPLAYED AS A CONDITION OF THE LICENCE GRANT FOR THIS SOFTWARE. ALL DERIVATEIVES WORKS MUST CARRY THIS NOTICE -->
     <!-- This is version of HBMonitor SP2ONG 2019-2025 -->
     <!-- HBlink Docker Installer ShaYmez 2021-2025 -->
 </p>
-</div>
+</footer>